+++ ChangeLog.md
@@ -2,12 +2,13 @@
 
 ## ChangeLog
 
-### 2.1.1-rc.0
+### 2.2.0-rc.0
 
     Release date: 
 
     - Define new AccountsCore.C.Verbose.USERS constant to be verbose when the 'users' collection is instanciated
     - Update cleanup transformation to keep the source of authentication in the user document
+    - Define new 'pwix.AccountsCore.m.sendVerificationEmail()' method
 
 ### 2.1.0